+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/Changelog.rst
@@ -16,7 +16,8 @@ Changed
 *******
 - Move PARTY_SCENE_ID to PartyConstants to enable other plugins to extend parties without hardcoding the party template id.
 - Use a byte array instead of a string for party custom data.
-Fixed
+- Allow a player to join the party game session even if it does not track players & teams.
+
 *****
 - Fixed an issue that would prevent `ResetPlayerReadyStateOn(ResetPlayerReadyStateMode.None)` from being applied correctly. 
 - Ensure that the stormancer user id is always used when sending a party invitation through Stormancer.